Two improvements:

- The RTP is now initialized the first time a file is missing. This way there is no startup delay, instead the delay is now when the first RTP asset is requested. Yes this only moves the problem but for some games this means 0 delay.
- lcf INI parser replaced with a simple parser that just reads all lines and skips until the Key/Value pair is found. No extra caching, that's why I don't consider this a fix for #1903, likely can be improved more.