@fmatthew5876 2019-09-25
> Delay the RTP detection until a file is missing.

How long does it take to do this routine the first time? 

I'm not sure I like this approach because it'll cause random stutter in the middle of game play the first time an asset is missing. That could be very jarring for the player.

Many games which don't use rtp still have missing assets. It's very common, and all those games will still pay the cost of initializing rtp.

I'd rather just pay the cost up front and deterministically, than have a random slowdown occur during game player.

> Registry Wine: Use custom INI parser (faster)

This looks like a step in the right direction. Ultimately we could just refactor it further to be most optimal. 

* Keep a set of all the keys we'll ever need in some array, tag each one with an enum. 
* Change `Registry` to an object, in it's constructor read the file and load the full set of keys once into an array of string values.
* Provide get methods to just return the key values by enum as an index into the array.

The result would be exactly one time reading the registry file and simple array indexing to retrieve values.