@Ghabry 2019-09-26
thanks for the feedback. Will work on a better solution.

When the RTP detect is fast enough there is no reason to use the "delay rtp detection" code.