@fmatthew5876 2019-11-18
With the delays added in #1940, maybe there is a possibility to use them to our advantage.

In most cases, I would expect the title screen, system graphic, and title music to be included in the game assets and not be RTP. The face graphics in the load game menu are likely to be custom for most serious games also.

Given that, how about we do this:

1. Load all assets needed for title and load screen only. 
    * If none of them in RTP, skip RTP detection. 
    * If any are RTP, then we have to wait for RTP detection on startup.
2. Upon loading the title screen, immediately kick off a background thread to start detecting RTP. 
3. Once we new game / load game and wait for transition out and the 1 second delay, block on black screen if RTP detection still hasn't finished.

With this approach we could achieve parallel and deterministic loading of RTP.