@fmatthew5876 2019-12-15
Can we push back the "delay rtp detection" piece to a follow up PR? The custom parser your wrote for wine results in a significant speedup and should go in sooner

Master
```
--------------------------------------------------                                                                                                                           
Benchmark           Time           CPU Iterations                                  
--------------------------------------------------                                                                                                                           
BM_InitRtp   31469504 ns   31464948 ns         22   
```

With new parser
```
--------------------------------------------------                                                                                                                            
Benchmark           Time           CPU Iterations                                                                                                                             
--------------------------------------------------                                                                                                                            
BM_InitRtp    5514361 ns    5512936 ns        137  
```