@Ghabry 2020-02-17
Using a Lowercase func that works in-place gives a 25% speedup and completely removing the Lowercasing (incompatible with Windows registry behaviour) gives another 75% speedup.

```
With normal LowerCase:
-----------------------------------------------------
Benchmark           Time             CPU   Iterations
-----------------------------------------------------
BM_InitRtp   38521503 ns     38474798 ns           18

With in-place Lowercase func
-----------------------------------------------------
Benchmark           Time             CPU   Iterations
-----------------------------------------------------
BM_InitRtp   25587533 ns     25436706 ns           27

Without Lowercase:
-----------------------------------------------------
Benchmark           Time             CPU   Iterations
-----------------------------------------------------
BM_InitRtp    7270756 ns      7258756 ns           95
```