@carstene1ns 2020-02-20
The last commit "lowercasing all in place" is wrong.
You lowercase the actual path. Also, you do it twice.

```
Debug: RTP path c:\\users\\carsten\\application data\\kadokawa\\common\\rpg maker 2003 rtp is invalid, not adding
```

```ini
[Software\\Wow6432Node\\KADOKAWA\\rpg2003] 1576949727
#time=1d5b82508a24236
"RuntimePackagePath"="C:\\users\\carsten\\Application Data\\KADOKAWA\\Common\\RPG Maker 2003 RTP"

```