@Ghabry 2020-02-20
ups. Removed the in-place lowercasing when inside of the section, thx. Won't hurt the parsing speed much as the section-searching is the slow part.